If we want to process data without estimating the PSF, use the guessed initial psf if we're unable to read an updated one from disk.
We should probably add a config option to control whether we even try to
read the old one.  This should be part of the rethink that means that the
code tries to continue when you've disabled something upstream rather than
aborting (e.g. aperture corrections if you've disabled PSF measurements)