Add task to select images from command-line, with associated plumbing (#2674).
WcsSelectImagesTask selects images by comparing their Wcs with the provided
patch coordinates.  This replaces BadSelectImagesTask as the default, as it's
much more useful.

Wcses are read at parse-time, so that when multiprocessing they are only read
once.  They are passed to the appropriate coadd Task, and then into the
WcsSelectImagesTask.  Existing SelectImagesTasks should not need to be updated,
since they do not implement the method whose signature has changed.  The
behaviour of the BaseSelectImagesTask when data identifiers are specified on
the command-line is to further restrict whatever the (presumably)
database-based SelectImagesTask returns, giving the user a bit more control.

Removed database-specific elements from BaseExposureInfo (assuming an
implementation is not appropriate for a 'Base' class; the _ind and _nextInd
members are for tracking values in a list of columns from a database query,
and getColumnNames() is clearly database-specific and unused in pipe_tasks).