Getting the new selection task ready to deliver.
Putting in the plumbing so we can get data to it from the command-line.
Reworked ButlerSelectImagesTask to use Wcs calculated before the Task code
starts, so that when using multiprocessing, the headers only have to be
read once.  Renamed ButlerSelectImagesTask to WcsSelectImagesTask, since
that's what it uses, rather than the butler.
Removed BadSelectImagesTask, as it's no longer required:
WcsSelectImagesTask is a suitable default.