Move PSF-matching config options into CoaddBase.
AssembleCoadd will need access to these options so it can install the
correct PSF into the final Exposure.

Also moved the CoaddBase.makeModelPsf function to an apply() method
on the config itself; in the future we'll probably want to make a registry for
model PSFs, and this is a step towards that (#2781).