calibrate: make astrometry failures non-fatal
In many cases, we can limp along without astrometry succeeding, so
there's no need to give up completely.  In some cases, we want to
ensure that we succeed, so added config variable 'requireAstrometry'
to allow a RumtimeError in astrometry to be fatal.

Conflicts:
	python/lsst/pipe/tasks/calibrate.py

This is essentially a cherry-pick of HSC's standalone commit:
https://github.com/HyperSuprime-Cam/pipe_tasks/commit/e9db5c0dcdca20e8f7ba71f24f8b797e71699352
with fixup:
https://github.com/HyperSuprime-Cam/pipe_tasks/commit/c2d89396923f9d589822c043ed8753647e70f3f6