WcsSelectImagesTask: use polygons for more precise selection (#2944)
We use lsst.geom.SphericalConvexPolygon to determine overlap between a
patch and images.  This is much more precise than using bounding boxes,
which means that we don't waste time reading images that don't overlap
the patch of interest.

Also added test of WcsSelectImagesTask and the use of its interface
in CoaddBaseTask.selectExposures.