measurePSF: Allow for option to reserve PSF candidates from fitting for cross-validation
To get a unique random seed for each CCD this required adding expId method for processImage

Conflicts:
	python/lsst/pipe/tasks/calibrate.py
	python/lsst/pipe/tasks/measurePsf.py
	python/lsst/pipe/tasks/processCoadd.py
	python/lsst/pipe/tasks/processImage.py
This is essentially a cherry-pick of HSC's HSc-966.
Note: 1) HSC-966 defined method getExpId which replicates the functionality of getExposureId.
Hence, getExpId was not included and all calls of getExpId were replaced with getExposureId.
2) DM-3693 implemented HSC's standalone ticket HyperSuprime-Cam/pipe_tasks@08a8ec0 that moved
CalibrateTask from ProcessImageTask to ProcessCcdTask. Therefore, getExpId is no longer
called from processImageTask. Instead, getExposureId is called from ProcessCcdTask and the
exposure id is passed to CalibrateTask.
3) HSC-966 sets the flag 'calib.psf.reserved' for all sources that were reserved from psf
fitting. This flag is named using the old "version 0" table schema approach. It has been
changed to 'calib_psfReserved' in compliance with the "version 1" table schema naming rule.