move CalibrateTask from ProcessImageTask into ProcessCcdTask
CalibrateTask seems to have been included in ProcessImageTask before the
days of CoaddPsf, when we needed to be able to generate a Psf for a coadd
using the coadd; this is no longer necessary, and its presence in
ProcessImageTask stands in the way of further refactoring to support the
plan for coadd multi-band deblending and measurement.

In order to implement matching between the detected sources and the
reference catalog, I've added an AstrometryTask to ProcessCoaddTask, since
it can no longer rely on the one in CalibrateTask.  This change does
change the config hierarchy, and so will break use of processCcd on
data repositories with existing data (if not using "--clobber-config").

Conflicts:
	python/lsst/pipe/tasks/calibrate.py
	python/lsst/pipe/tasks/processCcd.py
	python/lsst/pipe/tasks/processCoadd.py
	python/lsst/pipe/tasks/processImage.py

This is essentially a cherry-pick of HSC's standalone commit:
https://github.com/HyperSuprime-Cam/pipe_tasks/commit/08a8ec029dd52ac55e47b707a6905df061a40506

Note: this commit includes the ability to pass a schema into the
CalibrateTask and ProcessImageTask initializations.  While there
is no exisitng code that explicity passes a schema to these tasks,
this is considered a useful feature, so is being included (here
rather than in a separate feature commit to preserve the history
of its introduction into the HSC stack).