Add support for injecting fake sources in processCcd
This commit adds a new subtask interface for adding fake sources
in fakes.py, and a call to that interface in ProcessCcdTask which
creates a fakes subtask. This subtask will only run if doFakes
option is set to true, and the task is retargeted as the default
base fake algorithm is an abstract base class and can not be
initialized. Third-party code can retarget the 'fakes' subtask
to add their own simulation code into the pipeline.

Port note:
This came from hsc-166

fixup for fakes