calibrate: allow photometric calibration to be required for success
If requirePhotoCal is set, then raise an exception if the photocal
doesn't succeed; otherwise, it's OK to push on without it.
This only applies to the second, final photocal.  The initial photocal
may fail due to, e.g., a guess PSF that's nothing like the actual PSF
(and therefore star/galaxy separation is garbage), in which case the
second and final photocal (after having measured with the actual PSF)
should be successful.  It's the success of the second and final photocal
that we ultimately care about.