Need to reset ProcessCcdTask's schema to include fields propagated from calibrate
SchemaMapper doesn't modify the output Schema it's given in-place - it makes a
copy and modifies that.  Because ProcessImageTask was looking at self.schema,
which did not include the flag fields propagated from output fields, when we were
setting those we'd be mucking up the data in the next record (and if we ever tried
to do this on the last record, we'd could get a segfault).