Update tests/testProcessCcd.py as per the review
Make the tests more lenient, so they will be more robust against changes in algorithms.
Also use the butler to unpersist data where possible, as this is more robust than
assuming internals of the values returned by the task and tests persistence round trip.
In fact this showed a bug in persisting background lists, for which I am about to file a ticket.