Fix testProcessCcd.py to pass for old numpy
It turns out that older versions of numpy (1.7.1 at least and
maybe some newer than that) did not use double precision
aggregators on single precision arrays.  That doesn't provide
enough precision to pass the tests.  By requesting the
output be float64 one can force the aggregation in double
precision.  This fixes the test.