multiBand: allow schema to be specified in constructor
When these Tasks are run in their own stand-alone scripts, the schema
is loaded through the butler.  However, when they are run as part of a
common operation, the Tasks are instantiated before the schemas are
written out, and so they need to be given the schemas at construction
time.  This change makes both possible: the schema is used if provided,
otherwise loaded from the butler.

Conflicts:
	python/lsst/pipe/tasks/forcedPhotImage.py
	python/lsst/pipe/tasks/references.py