Remove default implementation from MergeSourceTask
MergeSourcesTask's implementation was getting in the way of its
subclasses a bit, by adding a field to the schema that they would not
ultimately use.  Since we're about to add a real implementation for
MergeMeasurementTask, we can now remove the old default implementation
without breaking the pipeline (at least not for very long).

Fixup for removal for MergeSources default impl