Stack: add option to do safe clipping when building coadd.
Identify objects in the tempExp files that need to be clipped by
looking at the difference between a clipped and unclipped coadd.  The clipped
masks that get created for each tempExp are not currently stored.

PORT: This code has moved locations from HscPipe to pipe_tasks in LSST.
Original commit can be found at
https://hsc-jira.astro.princeton.edu/jira/browse/HSC-1166