Remove non-positive variance pixels.
If the variance is negative, we can get exceptions from taking the sqrt.
If the variance is zero, the S/N is infinite and the pixel will be
identified as a source.

This is a workaround for issue DM-3201 ("Interpolation allows
non-positive variance"), cherry-picked from c42c0b99 on hscPipe.