calibrate: add curve of growth
The curve of growth corrects fluxes to a very large aperture.  The idea is
that the aperture correction corrects fluxes to a modest aperture with
spatial variation, while the curve of growth corrects to a very large
aperture ("infinity") with a single value for the entire CCD.  The
aperture correction radius (the calibFlux slot) should therefore be set
to the smallest radius outside which the correction does not vary
spatially.

We correct the flux in the calibFlux slot, and then the aperture correction
will correct everything else.  We also correct the multiple aperture
fluxes.