CalibrateTask: fix application of curve of growth
The CurveOfGrowth.getRatio method returns the ratio of small/large.
To correct the small aperture to the large aperture, we should therefore
divide rather than multiply.
Fixed the propagation of errors as well (was propagating as for addition
instead of multiplication or division).

Thanks to Bob Armstrong, who originally noticed this bug.