Update colorterms to support libraries and data from configs
Adapt the HSC version of colorterms, which adds:
- The ability to load data directly from config override files (instead of
requiring that obs_* packages contain data in library modules)
- Libraries of color terms, so one camera can support many different
kinds of reference catalog.
I adapted the code, eliminating reliance on eups and the assumption that that we use astrometry_net
indices for our reference catalogs. As a result one must specify a reference catalog name
in PhotoCalTask's config (the default is "*" which only works if your colorterm data is for a
single reference catalog). I also simplified the classes, e.g. combining ColortermConfig into Colorterm.