multiband: add warning about large footprints ignored by deblender
This is part of a workaround for large deblender footprints causing large
memory use and/or very slow processing.  We refuse to deblend those
footprints when running on a cluster and return to reprocess on a machine
with larger memory or more time if we consider those footprints important
to recover.  This warning alerts the user to the need to perform the
reprocessing.