Improve documentation and default config for PhotoCalTask colorterms
PhotoCalTask's config for colorterms had an unusable default for refCatName
(it was "*", but that value should not be a glob expression).
Changed to allow None as a value for refCatName and changed the behavior
of the task to handle that sensibly.

Improve how reference catalog names are used in ColortermLibrary