multiband: DetectCoaddSourcesTask output calexp_det exposure
As per
https://confluence.lsstcorp.org/display/DM/S15+Multi-Band+Coadd+Processing+Prototype
the DetectCoaddSourcesTask should output a calexp exposure with
updated background and mask plane information.  This was not being written,
so this updates the task to write out the calibrated exposure.  Note also
that the name has been changed to coaddName + "calexp_det" to distinguish
it from the "calexp" images output from processCoadd.py.