multiband: Set doMatchSources default to False
While the output from writeMatches is useful for debugging,
it is not required.  The default is being set to zero as, in
its current configuration, it does not work as the astrometry
task would have to be initiated (i.e. run) in order for the
matcher to be instantiated.  Changing the default back to
True awaits a refactoring of the astrometry matching API.