Detect if aperture correction cannot be measured due to no calib_psfUsed flag
In pipe_tasks CalibrateTask, by default aperture correction uses source flag "calib_psfUsed"
to decide if a source is acceptable to use for measuring aperture correction. If PSF determination
is disabled then this flag is never set and aperture correction will fail with a complaint
that there are 0 sources (which is true, but does not give a good indication of what's wrong).
I enhanced CalibrateTask.validate() to detect this situation: if doMeasureApCorr true
and measureApCorr.inputFilterFlag == "calib_psfUsed" and doPsf false then the config is invalid.