PropagateVisitFlagsTask: guard against non-contiguous ccdInputs
It is safer to read the ccdInputs value by value instead of by column,
since the latter way fails if the Catalog is non-contiguous. Since we
don't expect that this catalog is large, this should not have too bad
an effect on the performance in comparison with the other operations.