Replace meaningless/unusual checks in PhotoCal test.
There's no reason to put a threshold on the mean residuals in a fit,
because that doesn't reflect the quality of the fit (in fact, if it's
small that's probably an indication of overfitting).

In addition, the mean absolute residuals is a pretty unusual
statistic.

Instead, we now just report and threshold on RMS residuals and median
absolute residuals.