Prevent MakeDiscreteSkyMap from saving metadata
The metadata is not interesting, so don't save it.
I disabled it using the standard technique of having
MakeDiscreteSkyMap._getMetadataName() return None.
I left the code in the task runner that saves it,
because that task runner is more generally useful
and might be wanted for tasks that do save metadata.