Standardize the way MakeDiscreteSkyMapTask disables config saving
MakeDiscreteSkyMapTask was disabling saving of config by overriding
a method in its task runner, making the task runner less generally
useful and obscuring the fact that the task did not save its config.
I removed the override in the task runner and replaced it with
the standard means of disabling config saving: by having
the task's _getConfigName method return None.