Update testProcessCcd.py to be much pickier
Modified tests/testProcessCcd.py as follows:
- Compare values to 7 digits after the decimal point
(a fairly arbitrary choice), which is much pickier than before.
This is possible because updating Eigen fixed a fitting error
that gave different answers on Mac and linux, neither of
which was correct(DM-3699).
- Use exact comparison when comparing values determined from
the in-memory exposure and the unpersisted exposure.
- Print all values that are tested before testing any
(thus making it easier to determine new values when algorithms
change)