Update Config Parameters for Clipped Mask
Safe coadd clipping introduces a new mask plane, "CLIPPED". In order for
routines to see this plan and add appropriate pixel flags, we must specify
it as a plane to read. The configuration for pixelFlags algorithm has an
any and center plane, which specifies if the routine should look for the
plane anywhere in the foot print or near the center and set the flag
accordingly.