Added option to run legacy assembleCoadd task
Changed the "binary" command used to run assemble coadd. By default the
command will run the new SafeClipAssembleCoaddTask. If the --legacyCoadd
option is passed with the command, it will run the original coadd task,
and print a message warning the user to set configuration changes necessary
so future tasks run with a coadd made with the legacy code will not fail. In
the case where no arguments are passed, the default help message will be
triggered in the task argument parser. To ensure the extra
help is printed, the function needs to print its help when the
argument list has zero length.