Add a new task IngestCalibsTask for calibration images
It uses and expands IngestTask to parse through the file headers
for information and add to the calibration registry of the data
repository.  Dataset type (e.g. bias/dark/flat/fringe) is
determined by the header info OBSTYPE, and by default the
calibration images are valid on the observation date only.
Optional argument "validity" can be used to set a wider validity
period on the command line. Overlaps in validity ranges are fixed.