PhotoCal: add methods to optionally remove galaxies from matches
There is already a mechanism for extracting matches that are known
to be stars through the reference catalog.  This change allows us
to optionally use our own star/galaxy classifier.  This is optional
because the star/galaxy classification may be poor before a good
PSF has been measured.

This is essentially a cherry-pick of HSC-1086's:
https://github.com/HyperSuprime-Cam/meas_astrom/commit/78bcecd8faf590b4eda13813c5eac12bdcec3533#diff-caa9aa629cfed17b35838e3f0c4138d4R384

Note that photoCal.py lives here in pipe_tasks in the LSST stack
(as opposed to meas_astrom in the HSC stack).