Refactor and unify interpolation code.
Rework run function in interpImageTask to accept an image
of type maskedImage or Exposure and set psf appropriately.
Allow caller to supply either a mask plane OR a defects
list for interpolation.

Remove defect interpolation function from repairTask which
now creates interpImage as a subtask and calls its run
function for the defect interpolation.

Be more explicit and specific about how the fallbackValue for
edge interpolation is defined and computed in config parameters.