scaleVariance: use more robust algorithm
The former implementation could produce the wrong scaling in cases where the
image has strongly variable variance (e.g., 10 inputs contributed to half the
image, but only 1 input contributed to the other half) because it calculated
the variance of the image plane and the mean of the variance plane separately,
so that clipping can affect different pixels. Here, rather than clipping each
plane separately, we take a robust estimate of the standard deviation (based
on the IQR) after combining the two planes.

Moved the two implementations (in ProcessCoaddTask and DetectCoaddSourcesTask)
into a single function in coaddBase.