scaleVariance: use more robust algorithm
The former implementation could produce the wrong scaling in cases where
the image has strongly variable variance (e.g., 10 inputs contributed to
half the image, but only 1 input contributed to the other half) because
it calculated the variance of the image and the mean of the variance
separately so that clipping can affect different pixels.  Changed so
that the calculation is performed on a single image, and using a more
robust measure of the standard deviation (from the IQR).

Moved the two implementations (in ProcessCoaddTask and
DetectCoaddSourcesTask) into a single function in coaddBase.