CoaddDataIdContainer: use given dataset instead of deepCoadd
The use of deepCoadd (well, the "deep" part was configurable) was
hard-coded, but we now write background-subtracted coadds as
deepCoadd_calexp since HSC-1311.  Changed
CoaddDataIdContainer.makeDataRefList to respect the provided dataset
type instead of always grabbing deepCoadd.  However, the skymap
doesn't have the same flexibility, and is always deepCoadd_skyMap,
so simplified the parameter list.