makeCoaddTempExp: disable background restoration
Background restoration is used to create warps with a background in
them for the purpose of background matching.  However, background
matching is not currently working on pointed exposures (works OK
for drift scans), so we need background-subtracted warps to coadd
(or we're going to have lots of rejection and kill the S/N).