Update testCoadds.py unittest to DM-2915
As of DM-2915, the config setting:
self.measurement.plugins['base_PixelFlags'].masksFpAnywhere = ['CLIPPED']
is set as a default for MeasureMergedCoaddSourcesTask. However, this
CLIPPED mask plane only exists if a given coadd was created using the
newly implemented SafeClipAssembleCoaddTask. If a coadd was built using
AssembleCoaddTask, the CLIPPED mask plane is not present, so the above
default must be overridden to exclude it when using
MeasureMergedCoaddSourcesTask. This is the case for the mock coadd that
is assembled in the unittest code in testCoadds.py, so the config needs
to be set accordingly for the test to run properly.