Update documentation assemble method
This fixes a typo and documents that *args and **kwargs are passed but not used.  The parent task assumes arguments for the assemble method.  The overridden method is such that it doesn't need all those arguments, but it needs to be able to take them since the parent task calls this method.