Do not use full match list even if persisted
The schema aliasing gets too complicated if the full, unnormalized
match catalog is read in.  This is because only HSC persists this
catalog, so the HSC schemas are applied, but the LSST schemas are
the ones we need.  The alias mapping has been accommodated when
regenerating an unnormalized match list, so always read in the
normalized list and regenerate from it.