Fix logic for applying aperture corrections
With the current flow, the aperture corrections were being applied only
after all the measurement plugins had run through, independent of their
execution order.  This resulted in plugins whose measurements rely on
aperture corrected fluxes (i.e. with execution order > APCORR_ORDER)
being applied prior to the aperture correction, leading to erroneous
results.  The only plugin currently affected by this is
base_ClassificationExtendedness.

This fix indeed results in the correct order of operations, but it
is far from elegant and the logic of how and when aperture corrections
are applied should be reworked (on another ticket) to be less error-prone.