Make default dataset calexp
The default dataset was raw.  This normally isn't a problem, but
I was using a repository with only calexp data.  I expect that
makeDiscreteSkyMap will very frequently (always?) be run on
calexps since the whole point is to make a map that can be used
to change calexps into tempCoaddExps.