DetectAndMeasureTask: improve how aperture correction is measured
Measure aperture correction in CharacterizeImageTask (which has the
needed information, e.g. which sources were used to measure PSF).
Also perform a final measurement in that task after the PSF is measured
(like the original code). This eliminated the need for a callback function
in DetectAndMeasureTask, though I replaced that with a new flag "allowApCorr",
so that CharacterizeImageTask can detect and measure aperture correction
only after the PSF is determined.