fixes failing test
there was an issue where the CoaddsTestCase method getCalexpIds
was accessing a butler's mapper directly, but now there is a
repository class inbetween. Ideally this would be refactored to
acommodate multiple repositories but for the time being it is ok
to use the known state and assume exactly one repository.