Implement simple spacial indexer
This implements an indexer that provides the methods necessary to
shard a catalog and then look up the correct shard later.  The package
I use currently has a problem if you call the lookup_id method without
first calling another function.  This hack will go away if the package
is fixed or if we get an LSST maintained version.