Implement the reference catalog ingester
This implements the task to take files and parse them into
SourceCatalogs.  Some of the tasks implemented here are:
read the file into a numpy array (the whole file needs to fit
in memory), construct the SourceCatalog schema from the array
dtype, index the catalog and fill the SourceCatalog.

Command line tasks need a command runner.  Since this task
doesn't need to parse an `--id` argument, we an forego that
step.  Essentially, this just calls the task run method and
gives a place to persist the config.