Implement the catalog loader
This involves reconstituting the task that ingested the files.

The only config is the name of the config to read to create the
ingester task.  The loader then borrows the indexer from the
ingester task to grab the files from the repo it needs to load the
reference catalog.  The shards are only clipped if they land on the
boundary of the circular aperture.

This is a minimal override of the reference catalog loader.  It may
be possible to optimize by overriding other default implementations.