Include pixel padding to calls to createMatchMetadata
The createMatchMetadata function now allows for the pixelMargin to
be reflected in the radius set in the metadata by adding an optional
argument, border, to specify by how many pixels the bbox should be
grown before determining the radius.  This updates the calls to
this function to include the pixelMargin set in the config.