Convert run argument to config
Having default arguments in a run method in a task is not really
very useful as they can't be changed at runtime.  I modified the
argument that selects whether to use the extendedness value in choosing
photometric calibration sources to a config field so it can be
configured like the rest of the options.