multiband: use new MatchTask
For measureCoaddSources, it's helpful (for QA especially) to match
our catalog against a reference catalog. We were using an AstrometryTask
for this, but this isn't the right choice because they can allow
rotation and translation, filtering of our catalog, and they raise an
exception if the matching fails. And if you accidentally allow it to
fit a Wcs (which we have done in the past!), it can cause trouble for
your catalog. We just want a simple matching against the reference
catalog, which is provided by the new MatchTask.