Default local background to false for safe clipping
The safe clipping algorithm was designed and tuned with no extra background
estimations, thus doTempLocalBackground should default to False for that
task.  Additionally, the safe clipping algorithm reverts to the footprints
from single frame processing for very large footprints, thus we must also
turn it off in the single frame processing in order to get the desired safe
clipping behavior.