Fix problem with cov mulipliers
The covariance multipliers were being computed based on the initial
image in a series of images to be warped. This was causing problems
when the warp kernel size for a single image differed from that
computed for the first image. This commit fixes this problem by
warping all the images initially in order to determine the maximum
value required for the covariance multiplier and then re-doing the
warping with that determined multiplier value.