makeSkyMap: persist versions
Config and metadata persistence was disabled through the TaskRunner,
which is not the recommended way, and for this reason when version
peristence was added it to the TaskRunner, it didn't run here either.
Moved disabling of config and metadata persistence to
Task._getConfigName and Task._getMetadataName (the recommended way)
which will allow the version persistence to run.