ingestCalibs: change default mode to 'move'
The standard pattern will be that calibs are created outside the main
calib directory (e.g., in the rerun) so that they are not immediately
available for use by active processes. Then they need to be transferred
into the calib directory; 'move' is better than 'link' because we will
want to blow away the directory where they were originally created.