Force testCoadds to run in separate pytest process
This test creates and configures a single shared butler repository.
It takes 200MB and 25 seconds to create it and for the tests to
run in parallel each sub process would need that amount of disk
space in a temporary directory. This is possible but is much
slower than forced single-process execution.