Changes to save nImage mask in AssembleCoaddTask
As an option, allow assembleCoadd to save an ImageU named _nImage,
which contains the number of exposures used to create the stack for
each pixel.  Dataset is named type_nImage, where type is the type of
the coadd (e.g., deepCoadd_nImage).

Note that this image does not exclude pixels which are clipped by
the statisticsStack algorithm, only those masked by badMaskPlanes.