Move the setup code into setUpClass
setup_module runs when the test module is set up and it happens
even if a test class has been selected that does not need any
of that code execute. In particular, only one test class in this
file needs the data created in setup so the code that generates
it should move to the class setup.

This change is required to allow --dist=loadscope to work correctly
with pytest, otherwise the other two test classes trigger the set
up even if they are distributed to separate nodes.