refactor PhotoCalTask to make selection of sources easier
We want to be able to select sources within the reference catalog
as well as science sources. The DirectMatchTask allows us to do that,
so we'll use that instead of inheriting from a matching Task (and
we already have an astrometric solution, so that's appropriate).

Part of this change involves using the new ReserveSourcesTask
(to consolidate code); this is bound up with the changes to selection.