Replace doOverwrite option with new ctor argument.
Because we haven't added a new command-line option to set the ctor
argument, it's now impossible to achieve set reuse=True from the
command-line when running makeCoaddTempExp directly.  But it will be
possible from coaddDriver.py, and we expect that to be used instead
whenever the user might want reuse=True.