AssembleCoaddTask: set EDGE bits as SENSOR_EDGE
The EDGE pixels from the individual CCDs have printed through into the
coadd, but EDGE means "we couldn't search for sources in this area
because we couldn't convolve by the PSF near the edge of the image",
so this mask plane needs to be converted to something else if we want
to keep them. We do want to be able to identify pixels near the edge
of the detector because they will have an inexact CoaddPsf. We
rename EDGE pixels as SENSOR_EDGE.