assembleCoadd: set INEXACT_PSF mask plane
If any of the input images isn't represented in the coadd
(due to clipped pixels or chip gaps), the `CoaddPsf` will be
inexact. Flag these pixels.

Using a mask plane for this allows us to keep SENSOR_EDGE
and CLIPPED separate (in case the distinction is important)
while providing a convenient shorthand with a clear name that
covers both.