refactor scaleVariance into Task
We want to add a small-scale background subtraction (to remove any
extended features which will throw off the chi distribution of the
background) and detect excessively large values (which signal
incorrect results that may adversely affect downstream processing),
so this is getting more complicated and belongs in a Task so it can
be properly configured. The scaleVariance function now forwards to
the new Task.

Included in guarding against excessively large values is a fallback
method for calculating the rescaling factor using the statistics of
the image and variance planes separately; maybe it works in regimes
where the current method does not.