Record PSF-Matched ValidPolygons in CompareWarp
CompareWarp uses temporal information from PSF-matched warps
that have been shrunk by half the matching kernel. Only pixels in the
direct Warp that have coverage in the PSF-matched warps are coadded.
The CoaddPsf should therefore use these smaller polygons for computing
the visits that have contributed to the PSF. This change also creates
config in coaddBase 'matchingKernelSize' that controls the matchingKernel
in PSF-matching in makeCoaddTempExp and the number of pixels to shrink
the valid polygons in CompareWarpAssembleCoadd.