Catch and handle error in scaleVariance
When an image contains no good data, the scale varinace task would
raise an error that was not caught or handled anywhere. This change
simply catches the error, and returens a variance scaling of 1. This
does not handle the issue of bad data, but it does prevent the scale
variance task from bringing down processing, so that a task down the
line can handle any bad data.