Use getScalar and getArray instead of get
Use `getScalar` or `getArray` instead of deprecated `get`
on `PropertySet` and `PropertyList` instances.

Note one special case: GetRepositoryDataTask.getMetadataItems
used `get` and so could return scalars for some entries.
We are not using this method anywhere in the DM stack so I took the
liberty of changing it to use `asArray` and always returning array
values and documenting that.