Propagate photom/astrom calib flags to coadds
DM-4878 added a mechanism for propagating flags from the input visit
catalogues to the coadd catalogue indicating stars that were used
for measuring the PSF. Since then, DM-9050 added similar calibration
flags for astrometry and photometry which are here added to the list
of flags that get propagated from visits to coadds.

A test for the propagation of these flags is in ci_hsc.