Fix setting of calib_psf_candidate flag
The docstring for this flag reads:

"Flag set if the source was a candidate for PSF determination as
determined by the star selector."

However, if the reserve fraction was set to a non-zero value, an object
that was selected as a PSF candidate by the star selector but then
flagged as reserved would have the calib_psf_candidate flag set to False.
This is inconsistent with the docstring (and erroneously implies the
object was not deemed a suitable candidate). This updates the setting of
the flag to reflect its docstring description.

In an effort to help avoid this confusion again, this also makes the
variable name chage:

     psfCandidateList ==> psfDeterminerList

as this represents the list of candidate psf objects that are actually
being sent to the determiner.  It is a sub-selection of the full list
of potential psf candidates (as selected by the makePsfCandidates Task
and stored here in selectionResult.psfCandidates) omitting those (if
any) marked as reserved.