Don't assume Camera iteration matches ccdExposureId ordering.
The "observations" catalog used in the coadd tests needs to be
sorted for find() to work, but we never actually guaranteed this
in the past; the order was set by Camera iteration, which is
actually just dict iteration and hence has no guaranteed order.
This broke on DM-15528 when moving the Camera from Python
(dict) to C++ (unordered_map) changed that iteration order.