Refactor colorterm calculations
The existing code made it very painful to compute the colorterm corrections.
Now `Colorterm.getCorrectedMagnitudes()` does all the necessary work.

I am skeptical of the magnitude error calculations here, but I'm just copying
them from PhotoCalTask, so they're at least not any more wrong than before.

This change does fix a serious bug in the refMagErr calculation: it was
previously being done using the secondary flux, not the primary flux (the loop
was overwriting `refFluxArr` before it was used to calculate the refMagErr).