objectMasks: update parser to handle Goulding mask format
Andy Goulding has provided new masks based on Gaia DR2, and with a
better tuned size; but the format is subtly different:
* The rotation angle of a 'box' is not specified (it’s supposed to be
  zero anyway).
* The 'ID:' string is replaced with 'Gaia DR2'.
* Floating point values use scientific notation in places.

We've decided that (apart from the lack of "ID:", which we want in
order to be explicit) it's best to update the parser rather than fix
the mask files: the differences from what we have been expecting are
reasonable, and fixing the parser now means we may not have to fix
it much for the next version.

Expanded the regex to support the new format, defaulting the
rotation angle of boxes to 0. Added comments to the regex so it's
easier to understand.