First go at ingesting defects as real butler datasets
Note I had to call them "deflects" so the mapper special casing didn't
shadow my attempts.

Also, this uses a put to persist the defects to avoid an intermediate file.
This is not the standard practice as usually calibs are on disk and are simply
linked into a data repository.