measureCoaddSources: fix error from empty matches
We can't generate a proper denormalised match list from no matches,
because we don't have the schemas (which come from the matches);
so lsst.meas.astrom.denormalizeMatches raises RuntimeError if
there are no matches. We need to catch this case before calling
denormalizeMatches. We generate a dummy denormalized matches file
to keep the Gen3 butler happy (it's expecting an output of some
kind).