Add extra config for given camera Focal Plane unit
As noted in FocalPlaneBackgroundConfig() "the focal plane frame is usually
defined in units of microns or millimetres rather than pixels. As such,
their values will need to be revised according to each particular camera."
The algorithm here was developed and battle-tested against HSC in the
context of a camera focal plane defined in "pixels", but it is meant to
accommodate FPs defined in other units.  It turns out one place -- the
setting of the threshold for the minimum number of pixels contributing
for a bin to be considered "good" -- was failing to accommodate FP units
other than "pixel".  This accommodation is made here by adding a config
defining the size of a detector pixel in the units for which the given
camera geoms focal plane is defined.