Use case insensitive instrument consistency check
We tend to use camelCase for directory names but we want the
name of the instrument in the ecsv defect files to match the
name of the instrument reported by the translator and/or
INSTRUME header.

It seems highly unlikely that we would think it's an error
for a LATISS observation to be in a latiss directory so this
change lower cases the instrument names before comparing.