Replace doApplyUberCal with doApplyExternal configs.
The doApplyUberCal config has been deprecated in favor of
doApplyExternalPhotoCalib and doApplyExternalSkyWcs.  These names are
consistent with the updates to validate_drp in DM-21950, and allow different
external calibrations (jointcal, fgcmcal, fgcmcal_tract) to be used, and for
external photometric and astrometric calibrations to be applied separately.
Finally, useMeasMosaic has been properly deprecated.