Verify that Gen2 and Gen3 reference catalogs are consistent.
We're changing how we specify reference catalog names in configuration
in the transition from Gen2 to Gen3, and in the interim it seems
prudent to require that both [sets of] config options remain
consistent.  This of course will do nothing until we actually start
calling config.validate() in the Gen3 executors, but that's a separate
problem.