Grow artifacts by fewer pixels.
We were growing the artifacts too much w.r.t.
the footprints in the template. Shink nSigmaToGrow
on the warp detections to increase depth and produce
fewer stars with inexact_psf. Simultaneously, increase
the nSigmaToGrow on the static sky template to the
default 2.4 sigma. This better protects the cores of
bright stars from clipping.